refactor skill service helpers
